requirements.txt
bcrypt==3.2.0
blinker==1.4
Bootstrap-Flask==1.8.0
cffi==1.15.0
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Faker==9.8.3
flake8==4.0.1
flake8-import-order==0.18.1
flake8-print==4.0.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
Pillow==8.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.27
text-unidecode==1.3
Werkzeug==2.0.2
WTForms==3.0.0
